# api/openapi.yaml
...
openapi: "3.0.4"
info:
  title: Streamline API
  version: 0.1.0
  description: |
    REST API for Streamline, a unified media management platform that replaces
    the *arr stack (Radarr, Sonarr, Lidarr, Readarr) and Overseerr. Provides
    endpoints for managing movies, TV shows, indexers, download clients, media
    servers, quality profiles, user authentication, and media requests.
...
servers:
  - url: /api/v1
paths:
...
  /config/auth:
    get:
      operationId: getConfigAuth
      summary: Get auth configuration (admin)
      description: |
        Returns the current runtime-safe auth configuration. Changes made via
        PATCH take effect immediately — no restart required.
      tags: [config, admin]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          $ref: "#/components/responses/AuthConfig"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: updateConfigAuth
      summary: Patch auth configuration (admin)
      tags: [config, admin]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        $ref: "#/components/requestBodies/UpdateAuthConfig"
      responses:
        "200":
          $ref: "#/components/responses/AuthConfig"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key created via /users/me/api-keys
...
  requestBodies:
...
    UpdateAuthConfig:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthConfigPatch"
...
  responses:
...
    UnprocessableEntity:
      description: Validation or connection test failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    AuthConfig:
      description: Current runtime-safe auth configuration.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthConfigView"
...
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable, safe-to-display error message.
        code:
          type: string
          description: |
            Stable machine-readable error code (e.g. last_admin,
            self_delete_forbidden, email_exists). Present for domain errors
            where the caller needs to branch on the specific reason; absent
            for generic errors where the message is sufficient.
...
    AuthConfigView:
      type: object
      required: [registration_mode, session_ttl, oidc_default_role]
      properties:
        registration_mode:
          type: string
          enum: [disabled, open, invite]
        session_ttl:
          type: string
          description: Go duration string (e.g. "168h", "30m").
        oidc_default_role:
          type: string
          enum: [admin, member, request_only]
    AuthConfigPatch:
      type: object
      description: Only provided fields are applied.
      properties:
        registration_mode:
          type: string
          enum: [disabled, open, invite]
        session_ttl:
          type: string
          description: Go duration string (e.g. "168h", "30m").
        oidc_default_role:
          type: string
          enum: [admin, member, request_only]
...